Implement like unlike using javascript
Use remote links to handle unobtrusive javascript,
that allows the controller to respond to Ajax requests
