+++ app/controllers/likes_controller.rb
@@ -3,22 +3,18 @@ class LikesController < ApplicationController
 
   def create
     @like = current_user.likes.build(like_params)
-    if @like.save
-      flash[:success] = 'Thanks for liking!'
-    else
-      flash[:alert] = @like.errors.full_messages.join(', ')
+    flash[:alert] = @like.errors.full_messages.join(', ') unless @like.save
+    respond_to do |format|
+      format.js { render 'toggle.js.erb' }
     end
-    redirect_to root_path(anchor: "denounce_#{@like.likeable_id}")
   end
 
   def destroy
     @like = current_user.likes.find(params[:id])
-    if @like.destroy
-      flash[:success] = "Un like #{@like.likeable_type}!"
-    else
-      flash[:alert] = @like.errors.full_messages.join(', ')
+    flash[:alert] = @like.errors.full_messages.join(', ') unless @like.destroy
+    respond_to do |format|
+      format.js { render 'toggle.js.erb' }
     end
-    redirect_to root_path(anchor: "denounce_#{@like.likeable_id}")
   end
 
   protected
