+++ app/views/likes/toggle.js.erb
@@ -1,0 +1,10 @@
+
+function toggleLike() {
+  let likeId = '<%= @like.likeable_id %>'
+  let elementCountLike = document.querySelector(`#denounce-count-${likeId}`);
+  elementCountLike.previousElementSibling.remove()
+  let link = '<%= likeable?(@like.likeable) ? like_link_for(@like.likeable) : unlike_link_for(@like.likeable)%>'
+  elementCountLike.insertAdjacentHTML("beforebegin", link)
+  elementCountLike.innerText = '<%= @like.likeable.likes.count %>'
+}
+toggleLike()
